projects/info_board/requirements.txt
RPi.GPIO==0.7.0
luma.led-matrix==1.5.0
